+++ settings.gradle.kts
@@ -4,3 +4,4 @@ rootProject.name = "scalabot"
 include(":scalabot-app")
 include(":scalabot-extension")
 include("scalabot-ext-example")
+include("scalabot-meta")
